flask-sla/demo/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
pygments
pymysql